backend/backend/src/main/java/com/chess4everyone/backend/Ai_Model_Code/requirements.txt
# Chess4Everyone AI Model Code Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Chess
python-chess>=1.9.0

# ML Models
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
